+++ tests/tests_pytorch/accelerators/test_ipu.py
@@ -177,15 +177,20 @@ def test_optimization(tmpdir):
 def test_half_precision(tmpdir):
     class TestCallback(Callback):
         def setup(self, trainer: Trainer, pl_module: LightningModule, stage: str) -> None:
-            assert trainer.precision == "16"
+            assert trainer.precision == "16-mixed"
             raise SystemExit
 
     model = IPUModel()
     trainer = Trainer(
-        default_root_dir=tmpdir, fast_dev_run=True, accelerator="ipu", devices=1, precision=16, callbacks=TestCallback()
+        default_root_dir=tmpdir,
+        fast_dev_run=True,
+        accelerator="ipu",
+        devices=1,
+        precision="16-mixed",
+        callbacks=TestCallback(),
     )
     assert isinstance(trainer.strategy.precision_plugin, IPUPrecisionPlugin)
-    assert trainer.strategy.precision_plugin.precision == "16"
+    assert trainer.strategy.precision_plugin.precision == "16-mixed"
     with pytest.raises(SystemExit):
         trainer.fit(model)
 
@@ -194,7 +199,7 @@ def test_half_precision(tmpdir):
 def test_pure_half_precision(tmpdir):
     class TestCallback(Callback):
         def on_train_start(self, trainer: Trainer, pl_module: LightningModule) -> None:
-            assert trainer.strategy.precision_plugin.precision == "16"
+            assert trainer.strategy.precision_plugin.precision == "16-mixed"
             for param in trainer.strategy.model.parameters():
                 assert param.dtype == torch.float16
             raise SystemExit
@@ -202,22 +207,31 @@ def test_pure_half_precision(tmpdir):
     model = IPUModel()
     model = model.half()
     trainer = Trainer(
-        default_root_dir=tmpdir, fast_dev_run=True, accelerator="ipu", devices=1, precision=16, callbacks=TestCallback()
+        default_root_dir=tmpdir,
+        fast_dev_run=True,
+        accelerator="ipu",
+        devices=1,
+        precision="16-mixed",
+        callbacks=TestCallback(),
     )
 
     assert isinstance(trainer.strategy, IPUStrategy)
     assert isinstance(trainer.strategy.precision_plugin, IPUPrecisionPlugin)
-    assert trainer.strategy.precision_plugin.precision == "16"
+    assert trainer.strategy.precision_plugin.precision == "16-mixed"
 
     changed_dtypes = [torch.float, torch.float64]
     data = [torch.zeros((1), dtype=dtype) for dtype in changed_dtypes]
     new_data = trainer.strategy.batch_to_device(data)
-    assert all(val.dtype is torch.half for val in new_data)
+    assert all(val.dtype is torch.half for val in new_data), "".join(
+        [f"{dtype}: {val.dtype}" for dtype, val in zip(changed_dtypes, new_data)]
+    )
 
     not_changed_dtypes = [torch.uint8, torch.int8, torch.int32, torch.int64]
     data = [torch.zeros((1), dtype=dtype) for dtype in not_changed_dtypes]
     new_data = trainer.strategy.batch_to_device(data)
-    assert all(val.dtype is dtype for val, dtype in zip(new_data, not_changed_dtypes))
+    assert all(val.dtype is dtype for val, dtype in zip(new_data, not_changed_dtypes)), "".join(
+        [f"{dtype}: {val.dtype}" for dtype, val in zip(not_changed_dtypes, new_data)]
+    )
 
     with pytest.raises(SystemExit):
         trainer.fit(model)
@@ -531,8 +545,8 @@ def test_multi_optimizers_fails(tmpdir):
 def test_precision_plugin():
     """Ensure precision plugin value is set correctly."""
 
-    plugin = IPUPrecisionPlugin(precision=16)
-    assert plugin.precision == "16"
+    plugin = IPUPrecisionPlugin(precision="16-mixed")
+    assert plugin.precision == "16-mixed"
 
 
 @RunIf(ipu=True)